+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- DEPRECATED API `require("ergoterm.terminal")` use `require("ergoterm")` instead.
+
 ## 0.8.0 - 2025-10-15
 
 - BREAKING: `select()` & `select_started()` now take table of options instead of ordered parameters.